Fill out towards all edges
The behaviour before was to make sure it filled up 100% in one direction
but ensure that grid had an exakt 1:1 ratio in positions. This does
however cause a black border at the bottom or to the right if ratios
doesn't align perfectly.

This border should only be 1/2 the width of character at maximum, I
decided to make the horizontal and vertical positions independent of
each other. The extra pixels added between characters should not be
visible with the naked eye and the image is still visibly pleasing, now
completely without any wasted space at the edges. If this causes
problems in the future, a more intelligent "gridding" algorithm may be
needed.

This would be less of a problem if SDL_ttf allowed floating point
position and sizes.