merge with progress in Heisig's RTK
You can provide the highest kanji index
you reached in Heisig's
Remember the Kanji (RTK)
with
-i <index>
--heisig-index <index>

You can also set the --heisig-color,
which is green by default.

The heisig progress will only be displayed
when the kanji is unseen in WaniKani.
Otherwise the WaniKani color/status will be shown.

The kanji order was replaced with the
6th edition Heisig RTK order, 2200 kanji.
One additional kanji that is only in WaniKani
was added at the end.